+++ libparistraceroute/algorithms/outputs/.gitignore
@@ -1,0 +1,2 @@
+.deps/
+.dirstamp
