+++ libparistraceroute/containers/set.h
@@ -1,7 +1,8 @@
 #ifndef SET_H
 #define SET_H
 
-#include <stdbool.h>
+#include <stdbool.h>                // bool
+#include <stddef.h>                 // size_t
 #include "containers/object.h"
 
 typedef struct {
@@ -18,7 +19,7 @@ typedef struct {
  * \param element_free Callback used to free element (may be set to NULL).
  *    If NULL, the set does not release references that it contains.
  * \param element_dump Callback used to dump element (may be set to NULL).
- *    If NULL, the set cannot be printed using set_dump. 
+ *    If NULL, the set cannot be printed using set_dump.
  * \param element_compare Callback used to compare elements (mandatory).
  */
